requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
slack-sdk>=3.27.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
msal>=1.28.0
snowflake-connector-python>=3.8.0
simple-salesforce>=1.12.5
jinja2>=3.1.3
aiohttp>=3.9.0
pytz>=2024.1
